moved DOM calls of tab creationto tabs.js
