+++ src/css/homepage.css
@@ -76,7 +76,7 @@ span {
   width: 100%;
   text-align: center;
   list-style: none;
-  padding: 5px;
+  padding: 5px 0;
   margin: 10px 10%;
 }
